requirements.txt
pandas>=2.0
numpy>=1.24
requests>=2.31
pyyaml>=6.0
pyupbit>=0.2.34
feedparser>=6.0.10
matplotlib>=3.8
